Major refactoring
- exception handling
- partial dry application
- consistent package names (lowercase)
- remove unused classes
- fixes, mark pending fixes with FIXME
- @Autowired removed
- remove Travis
