--- .travis.yml
@@ -1,8 +1,0 @@
-sudo: false
-language: java
-jdk:
-  - oraclejdk9
-cache:
-  directories:
-    - $HOME/.m2
-script: sh -c 'cd backend && mvn clean package'
