--- backend/src/main/java/com/intive/shopme/UserRegistration/ValidInvoiceIfInvoiceRequestedValidator.java
@@ -1,30 +1,0 @@
-package com.intive.shopme.UserRegistration;
-
-import org.springframework.stereotype.Component;
-
-import javax.validation.ConstraintValidator;
-import javax.validation.ConstraintValidatorContext;
-
-@Component
-public class ValidInvoiceIfInvoiceRequestedValidator implements
-        ConstraintValidator<ValidInvoiceIfInvoiceRequested, Users> {
-
-    @Override
-    public void initialize(ValidInvoiceIfInvoiceRequested constraintAnnotation) {
-    }
-
-    @Override
-    public boolean isValid(Users value, ConstraintValidatorContext context) {
-        boolean result = true;
-        if (value.getInvoiceRequest()) {
-            Invoice invoice = value.getInvoice();
-            result = (invoice != null) && invoice.hasCompanyDetails();
-
-            if (result) {
-                Address invoiceAddress = invoice.getInvoiceAddress();
-                result &= (invoiceAddress != null) && invoiceAddress.isFilled();
-            }
-        }
-        return result;
-    }
-}
