+++ backend/src/main/java/com/intive/shopme/controller/CategoryController.java
@@ -6,7 +6,7 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiResponse;
 import io.swagger.annotations.ApiResponses;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -20,6 +20,7 @@ import java.util.UUID;
 
 import static com.intive.shopme.config.ApiUrl.CATEGORIES;
 
+@Validated
 @RestController
 @RequestMapping(value = CATEGORIES)
 @Api(value = "category", description = "REST API for categories operations", tags = "Categories")
@@ -27,7 +28,6 @@ public class CategoryController {
 
     private final CategoryService service;
 
-    @Autowired
     public CategoryController(CategoryService service) {
         this.service = service;
     }
@@ -39,13 +39,12 @@ public class CategoryController {
     }
 
     @ApiResponses(value = {
-            @ApiResponse(code = 404, message = "Category with this ID has not found"),
             @ApiResponse(code = 409, message = "Category already exist")
     })
     @PostMapping
     @ApiOperation(value = "Saves new category")
     public Category add(@RequestBody Category category) {
-        return service.add(category);
+        return service.create(category);
     }
 
     @DeleteMapping(value = "{id}")
