+++ backend/src/main/java/com/intive/shopme/model/Identifiable.java
@@ -1,8 +1,7 @@
-package com.intive.shopme.base;
+package com.intive.shopme.model;
 
 import io.swagger.annotations.ApiModelProperty;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.Data;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.GeneratedValue;
@@ -10,8 +9,7 @@ import javax.persistence.Id;
 import javax.persistence.MappedSuperclass;
 import java.util.UUID;
 
-@Getter
-@Setter
+@Data
 @MappedSuperclass
 public abstract class Identifiable {
 
@@ -24,5 +22,4 @@ public abstract class Identifiable {
     @ApiModelProperty(value = "Represents unique id number", position = 1,
             example = "5d214c01-95c3-4ec4-8f68-51dfb80b191c")
     private UUID id;
-
 }
