+++ backend/src/main/java/com/intive/shopme/model/Owner.java
@@ -1,6 +1,5 @@
 package com.intive.shopme.model;
 
-import com.intive.shopme.base.Identifiable;
 import io.swagger.annotations.ApiModel;
 import io.swagger.annotations.ApiModelProperty;
 import lombok.Data;
@@ -13,9 +12,7 @@ import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
-import static com.intive.shopme.config.AppConfig.USER_DESCRIPTION_MAX_LENGTH;
-import static com.intive.shopme.config.AppConfig.USER_NAME_MAX_LENGTH;
-import static com.intive.shopme.config.AppConfig.USER_NAME_MIN_LENGTH;
+import static com.intive.shopme.config.AppConfiguration.*;
 
 @Entity
 @ApiModel(value = "Offer's user", description = "Represents offer's owner")
