+++ backend/src/main/java/com/intive/shopme/model/exception/AlreadyExistException.java
@@ -1,0 +1,8 @@
+package com.intive.shopme.model.exception;
+
+public class AlreadyExistException extends RuntimeException {
+
+    public AlreadyExistException(String message) {
+        super(message);
+    }
+}
