+++ backend/src/main/java/com/intive/shopme/repository/CategoryRepository.java
@@ -9,7 +9,7 @@ import java.util.UUID;
 @Repository
 public interface CategoryRepository extends JpaRepository<Category, UUID> {
 
-    boolean existsByName(String name);
+    Category findByName(String name);
 
-    boolean existsByTranslateKey(String translateKey);
+    Category findByTranslateKey(String translateKey);
 }
