+++ backend/src/main/java/com/intive/shopme/service/CategoryService.java
@@ -1,23 +1,18 @@
 package com.intive.shopme.service;
 
 import com.intive.shopme.model.Category;
+import com.intive.shopme.model.exception.AlreadyExistException;
 import com.intive.shopme.repository.CategoryRepository;
-import com.intive.shopme.util.RepositoryVerifier;
-import com.intive.shopme.util.validation.error.AlreadyExistException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
-import static com.intive.shopme.config.AppConfig.CATEGORY_NOT_FOUND;
-
 @Service
-public class CategoryService {
+public class CategoryService extends Validated<Category> {
 
     private final CategoryRepository repository;
 
-    @Autowired
     public CategoryService(CategoryRepository repository) {
         this.repository = repository;
     }
@@ -26,18 +21,24 @@ public class CategoryService {
         return repository.findAll();
     }
 
-    public Category add(Category category) {
-        if (repository.existsByName(category.getName())) {
-            throw new AlreadyExistException("Category with this name already exist");
-        }
-        if (repository.existsByTranslateKey(category.getTranslateKey())) {
-            throw new AlreadyExistException("Category with this translateKey already exist");
-        }
+    public Category create(Category category) {
+        checkExistence(category);
+        validate(category);
         return repository.save(category);
     }
 
     public Category getCategoryById(UUID id) {
-        RepositoryVerifier.throwNotFoundExceptionIfEntityNotFound(id, repository, CATEGORY_NOT_FOUND);
         return repository.getOne(id);
+    }
+
+    private void checkExistence(Category category) {
+        var foundByName = repository.findByName(category.getName());
+        if (foundByName != null) {
+            throw new AlreadyExistException(foundByName + " - name already exist");
+        }
+        var foundByTranslateKey = repository.findByTranslateKey(category.getTranslateKey());
+        if (foundByTranslateKey != null) {
+            throw new AlreadyExistException(foundByTranslateKey + " - translate key already exist");
+        }
     }
 }
