+++ backend/src/main/java/com/intive/shopme/user/registration/UserRepository.java
@@ -1,0 +1,10 @@
+package com.intive.shopme.user.registration;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+import java.util.UUID;
+
+@Repository
+interface UserRepository extends JpaRepository<User, UUID> {
+}
