--- backend/src/main/java/com/intive/shopme/util/validation/error/ErrorResponse.java
@@ -1,11 +1,0 @@
-package com.intive.shopme.util.validation.error;
-
-import lombok.Data;
-import org.springframework.http.HttpStatus;
-
-@Data
-public class ErrorResponse {
-
-    private final HttpStatus httpStatus;
-    private final String message;
-}
