--- backend/src/main/java/com/intive/shopme/util/validation/error/NotFoundException.java
@@ -1,9 +1,0 @@
-package com.intive.shopme.util.validation.error;
-
-import lombok.Data;
-
-@Data
-public class NotFoundException extends RuntimeException {
-
-    private final String name;
-}
