+++ backend/src/main/java/com/intive/shopme/validation/CategoryValidator.java
@@ -1,15 +1,13 @@
-package com.intive.shopme.util.validation.validation;
+package com.intive.shopme.validation;
 
 import com.intive.shopme.model.Category;
 import com.intive.shopme.service.CategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
 import java.util.UUID;
 
-@Component
 public class CategoryValidator implements ConstraintValidator<CategoryCheck, Category> {
 
     @Autowired
