+++ backend/src/main/java/com/intive/shopme/validation/LinkInTextValidator.java
@@ -1,4 +1,4 @@
-package com.intive.shopme.util.validation.validation;
+package com.intive.shopme.validation;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
